more refactors 🙈
