--- src/components/families/family-form/index.ts
@@ -1,6 +1,0 @@
-export {
-  default,
-  familyResponseToFamilyFormData,
-  studentFormDataToStudentRequest,
-} from "./FamilyForm";
-export type { FamilyFormData } from "./FamilyForm";
