--- src/components/families/family-form/student-fields/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./StudentFields";
