+++ src/components/families/family-form/student-form/StudentForm.tsx
@@ -5,10 +5,16 @@ import { Add, RemoveCircle } from "@material-ui/icons";
 import { makeStyles } from "@material-ui/styles";
 
 import { StudentRequest } from "api/types";
+import Field from "components/common/field";
+import DefaultFieldKey from "constants/DefaultFieldKey";
+import { DefaultFields } from "constants/DefaultFields";
+import FieldVariant from "constants/FieldVariant";
 import StudentRole from "constants/StudentRole";
 import { DynamicField } from "types";
 
-import StudentFields from "../student-fields";
+// unique identifier for children form components
+let CHILD_KEY_COUNTER = 1;
+let GUEST_KEY_COUNTER = 1;
 
 const useStyles = makeStyles(() => ({
   addButton: {
@@ -38,26 +44,56 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
+export const generateKey = (
+  role: StudentRole.CHILD | StudentRole.GUEST
+): number => {
+  let key = 0;
+  if (role === StudentRole.CHILD) {
+    key = CHILD_KEY_COUNTER;
+    CHILD_KEY_COUNTER += 1;
+  } else if (role === StudentRole.GUEST) {
+    key = GUEST_KEY_COUNTER;
+    GUEST_KEY_COUNTER += 1;
+  }
+  return key;
+};
+
 export type StudentFormData = StudentRequest & { index: number };
+
+const defaultStudentData: StudentRequest = {
+  [DefaultFieldKey.FIRST_NAME]: "",
+  [DefaultFieldKey.LAST_NAME]: "",
+  [DefaultFieldKey.DATE_OF_BIRTH]: null,
+  information: {},
+};
 
 type Props = {
   dynamicFields: DynamicField[];
-  onAddStudent: () => void;
-  onDeleteStudent: (id: number) => void;
-  onUpdateStudent: (id: number, data: StudentRequest) => void;
+  onChange: (students: StudentFormData[]) => void;
   role: StudentRole.CHILD | StudentRole.GUEST;
   students: StudentFormData[];
 };
 
-const StudentForm = ({
-  dynamicFields,
-  onAddStudent,
-  onDeleteStudent,
-  onUpdateStudent,
-  role,
-  students,
-}: Props) => {
+const StudentForm = ({ dynamicFields, onChange, role, students }: Props) => {
   const classes = useStyles();
+
+  const onAddStudent = (): void => {
+    onChange([
+      ...students,
+      { ...defaultStudentData, index: generateKey(role) },
+    ]);
+  };
+
+  const onUpdateStudent = (index: number, data: StudentRequest): void => {
+    const studentsData = [...students];
+    studentsData[index] = { ...studentsData[index], ...data };
+    onChange([...studentsData]);
+  };
+
+  const onDeleteStudent = (index: number): void => {
+    onChange(students.filter((e) => e.index !== index));
+  };
+
   return (
     <Box display="flex" paddingBottom={6}>
       <Box className={classes.rowContainer}>
@@ -80,12 +116,51 @@ const StudentForm = ({
               )}
             </Box>
             <div>
-              <StudentFields
-                dynamicFields={dynamicFields}
-                onChange={(data) => onUpdateStudent(i, data)}
-                role={role}
-                student={student}
+              <Field
+                field={{ ...DefaultFields.FIRST_NAME, role }}
+                onChange={(value) =>
+                  onUpdateStudent(i, { ...student, first_name: value })
+                }
+                value={student.first_name}
+                variant={FieldVariant.COMPACT}
               />
+              <Field
+                field={{ ...DefaultFields.LAST_NAME, role }}
+                onChange={(value) =>
+                  onUpdateStudent(i, { ...student, last_name: value })
+                }
+                value={student.last_name}
+                variant={FieldVariant.COMPACT}
+              />
+              <Field
+                field={{ ...DefaultFields.DATE_OF_BIRTH, role }}
+                onChange={(value) => {
+                  const dob = value || null;
+                  onUpdateStudent(i, {
+                    ...student,
+                    date_of_birth: dob,
+                  });
+                }}
+                value={student.date_of_birth || ""}
+                variant={FieldVariant.COMPACT}
+              />
+              {dynamicFields.map((field) => (
+                <Field
+                  key={field.id}
+                  field={field}
+                  onChange={(value) =>
+                    onUpdateStudent(i, {
+                      ...student,
+                      information: {
+                        ...student.information,
+                        [field.id]: value,
+                      },
+                    })
+                  }
+                  value={student.information[field.id] ?? ""}
+                  variant={FieldVariant.COMPACT}
+                />
+              ))}
             </div>
           </Box>
         ))}
