+++ src/components/families/family-form/student-form/index.ts
@@ -1,2 +1,2 @@
-export { default } from "./StudentForm";
+export { default, generateKey } from "./StudentForm";
 export type { StudentFormData } from "./StudentForm";
